+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 This is the changelog of our "JTL WooCommerce Connector".
 
 ## Unreleased
+- CO-2514 - prevent duplicate product images 
+- CO-2394 - delete removed custom attributes
+- CO-2460 - implement compatibility with germanized pro food 
 - CO-2582 - merge wpml with wc connector
 
 ## 1.42.2 _2024-02-14_